VPLAY-10193 fix for dai2 test, moved mp4demux.cpp into mp4demux.hpp
Reason for Change: add'l cleanup and gst test harness regression
Risk: Low
